fix bug

+++ src/tabdpt/regressor.py
@@ -30,7 +30,7 @@ class TabDPTRegressor(TabDPTEstimator, RegressorMixin):
                 task=self.mode,
             )
             
-            return pred.float().squeeze().detach().cpu().numpy()
+            return pred.float().squeeze().detach().cpu().float().numpy()
         else:
             pred_list = []
             for b in range(math.ceil(len(self.X_test) / self.inf_batch_size)):
@@ -57,7 +57,7 @@ class TabDPTRegressor(TabDPTEstimator, RegressorMixin):
 
                 pred_list.append(pred)
 
-            return torch.cat(pred_list).squeeze().detach().cpu().numpy()
+            return torch.cat(pred_list).squeeze().detach().cpu().float().numpy()
 
     def _ensemble_predict(self, X: np.ndarray, n_ensembles: int, context_size: int = 128):
         logits_cumsum = None
